+++ data-raw/LinTests.R
@@ -1,5 +1,4 @@
-library('ape')
-library('TreeTools')
+library('TreeTools', quietly = TRUE, warn.conflicts = FALSE)
 library('TreeDist')
 library('TreeDistData')
 
@@ -14,7 +13,7 @@ message("Running tests on ", nTrees, ' ', nTip, "-leaf trees; ",
         replicates, " replicates.")
 
 LinTestOneSet <- function (nTip, k, nTrees) {
-  skeleton <- RandomTree(seq_len(k))
+  skeleton <- RandomTree(seq_len(k), root = TRUE)
   structure(lapply(seq_len(nTrees), function (XX) {
     tr <- skeleton
     for (i in k + seq_len(nTip - k))
@@ -110,7 +109,7 @@ compAllMethods <- c('rf', 'icrf',
                     'pid', 'msid', 'cid', 'qd', 'nye',
                     'ms', 'mast', 'masti',
                     'nni_l', 'nni_L', 'nni_U', 'nni_u', 'spr',
-                    'tbr_l', 'tbr_u', 'path')
+                    'tbr_l', 'tbr_u', 'path', 'es', 'kc')
 linTestReturn <- matrix(FALSE, nrow = length(compAllMethods), ncol = 5L,
                         dimnames = list(compAllMethods,
                                         c('spc', 'pam', 'h.cmp', 'h.sng', 'h.avg')))